requirements.txt
feedparser
requests
scikit-learn
python-dateutil